+++ src/z80asm2/CHANGES.md
@@ -4,6 +4,11 @@ Z88DK Z80 Module Assembler Change Log
 2016
 ----
 
+- 2016-07-28
+
+  Utility module and tests to read lines with any EOL terminator (CR, LF, CR-LF)
+  and search for a file in a lookup-path.
+
 - 2016-07-23
 
   Command line parser based on re2c. Parse option --help. Read file list